is the fix this easy?

+++ tinygrad/runtime/ops_gpu.py
@@ -76,7 +76,7 @@ class CLAllocator(LRUAllocator):
     check(cl.clEnqueueReadBuffer(self.device.queue, src, False, 0, len(dest)*dest.itemsize, from_mv(dest), 0, None, None))
     self.device.synchronize()
   def offset(self, buf, size:int, offset:int):
-    buf_info = cl.cl_buffer_region(origin=ctypes.c_uint64(offset), size=ctypes.c_uint64(size))
+    buf_info = cl.cl_buffer_region(origin=offset, size=size)
     return checked(cl.clCreateSubBuffer(buf, cl.CL_MEM_READ_WRITE,
                         cl.CL_BUFFER_CREATE_TYPE_REGION, ctypes.byref(buf_info), status := ctypes.c_int32()), status)
